Module jet_margin::cpi::accounts

An Anchor generated module, providing a set of structs mirroring the structs deriving Accounts, where each field is an AccountInfo. This is useful for CPI.

Structs

AccountingInvoke

Generated CPI struct of the accounts for AccountingInvoke.

AdapterInvoke

Generated CPI struct of the accounts for AdapterInvoke.

CloseAccount

Generated CPI struct of the accounts for CloseAccount.

ClosePosition

Generated CPI struct of the accounts for ClosePosition.

CreateAccount

Generated CPI struct of the accounts for CreateAccount.

LiquidateBegin

Generated CPI struct of the accounts for LiquidateBegin.

LiquidateEnd

Generated CPI struct of the accounts for LiquidateEnd.

LiquidatorInvoke

Generated CPI struct of the accounts for LiquidatorInvoke.

RefreshPositionMetadata

Generated CPI struct of the accounts for RefreshPositionMetadata.

RegisterPosition

Generated CPI struct of the accounts for RegisterPosition.

UpdatePositionBalance

Generated CPI struct of the accounts for UpdatePositionBalance.

VerifyHealthy

Generated CPI struct of the accounts for VerifyHealthy.