Module jet_margin::instruction

An Anchor generated module containing the program’s set of instructions, where each method handler in the #[program] mod is associated with a struct defining the input arguments to the method. These should be used directly, when one wants to serialize Anchor instruction data, for example, when speciying instructions on a client.

Modules

state

Instruction struct definitions for #[state] methods.

Structs

AccountingInvoke

Instruction.

AdapterInvoke

Instruction.

CloseAccount

Instruction.

ClosePosition

Instruction.

CreateAccount

Instruction.

LiquidateBegin

Instruction.

LiquidateEnd

Instruction.

LiquidatorInvoke

Instruction.

RefreshPositionMetadata

Instruction.

RegisterPosition

Instruction.

UpdatePositionBalance

Instruction.

VerifyHealthy

Instruction.