Module jet_margin::jet_margin

This crate documents the instructions used in the margin program of the jet-v2 repo. Handler functions are described for each instruction well as struct parameters (and their types and descriptions are listed) and any handler function parameters aside from parameters that exist in every instruction handler function. Parameters of events emitted for the purposes of data logging are also included.

Functions

accounting_invoke

Perform an action by invoking other programs, allowing them only to refresh the state of the margin account to be consistent with the actual underlying prices or positions, but not permitting new position changes.

adapter_invoke

Perform an action by invoking other programs, allowing them to alter the balances of the token accounts belonging to this margin account.

close_account

Close a user’s margin account

close_position

Close out a position, freeing up space in the account.

create_account

Create a new margin account for a user

liquidate_begin

Begin liquidating an account

liquidate_end

Stop liquidating an account

liquidator_invoke

Perform an action by invoking another program, for the purposes of liquidating a margin account.

refresh_position_metadata

Update the metadata for a position stored in the margin account, in the case where the metadata has changed after the position was created. This instruction does the following:

register_position

Register a position for some token that will be custodied by margin. Currently this applies to anything other than a claim.

update_position_balance

Update the balance of a position stored in the margin account to match the actual balance stored by the SPL token acount.

verify_healthy

Verify that the account is healthy, by validating the collateralization ration is above the minimum.