Function jet_margin::jet_margin::adapter_invoke

pub fn adapter_invoke<'info>(
    ctx: Context<'_, '_, '_, 'info, AdapterInvoke<'info>>, 
    data: Vec<u8>
) -> Result<()>

Perform an action by invoking other programs, allowing them to alter the balances of the token accounts belonging to this margin account.

/// ## adapter_invoke.rs

This instruction does the following:

If a read account has the liquidation parameter set to a pubkey:

a. This means that that margin account is already under liquidation by the liquidator at that pubkey.

b. Return ErrorCode::Liquidating.

Emit the AdapterInvokeBegin event for data logging (see table below).

Check if any positions that have changed via adapters.

a. For each changed position, emit each existing adapter position as an event (see table below).

Emit the AdapterInvokeEnd event for data logging (see table below).

Verify that margin accounts positions via adapter are healthy.

Return Ok(()).

Parameters of adapter_invoke.rs:

Name	Description

owner	The authority that owns the margin account.

margin_account	The margin account to proxy an action for.

adapter_program	The program to be invoked.

adapter_metadata	The metadata about the proxy program.

Events emitted by adapter_invoke.rs:

Event Name	Description

events::AdapterInvokeBegin	Marks the start of the adapter invocation (includes the margin account pubkey and the adapter program pubkey).

events::PositionEvent (Note that each single event represents a different adapter position)	Each adapter position is emitted as an event (includes the margin account, the adapter program, the accounts, and a value of true for the field signed.

events::AdapterInvokeEnd	Marks the ending of the adapter invocation (includes no data except for the event itself being emitted).