Function jet_margin::jet_margin::close_position

pub fn close_position(
    ctx: Context<'_, '_, '_, '_, ClosePosition<'_>>
) -> Result<()>

Close out a position, freeing up space in the account.

This instruction does the following:

Load the margin account.

Verify the authority of the margin account.

Unregister the existing position from the margin account.

If the token account authority is the same as the margin account authority, close the token account.

Emit the events::PositionClosed event (see Events table below) for data logging.

Return Ok(()).

Accounts expected with close_position.rs:

Name	Type	Description

authority	signer	The authority that can change the margin account.

receiver	writable	The receiver for the rent released.

margin_account	writable	The margin account with the position to close.

position_token_mint	read_only	The mint for the position token being deregistered.

token_account	writable	The token account for the position being closed.

token_program	read_only	The spl token program.

Events emitted by close_position.rs:

Event Name	Description

events::PositionClosed	The closed position.