Function jet_margin::jet_margin::liquidate_begin

pub fn liquidate_begin(
    ctx: Context<'_, '_, '_, '_, LiquidateBegin<'_>>
) -> Result<()>

Begin liquidating an account

The account will enter a state preventing the owner from taking any action, until the liquidator process is complete.

Requires the liquidator_metadata account, which restricts the signer to those approved by protocol governance.

Accounts

Name	Type	Description

margin_account	writable	The account in need of liquidation.

payer	signer	The address paying rent.

liquidator	signer	The liquidator account performing the liquidation.

liquidator_metadata	read_only	The metadata describing the liquidator.

liquidation	writable	The account to persist the state of liquidation.

system_program	read_only	The system native program.

Events

Event Name	Description

events::LiquidationBegun	The event marking the beginning of liquidation.