Function jet_margin::jet_margin::liquidate_end

pub fn liquidate_end(
    ctx: Context<'_, '_, '_, '_, LiquidateEnd<'_>>
) -> Result<()>

Stop liquidating an account

liquidate_end.rs

This instruction does the following:

Let account be a mutable reference to the margin account.

Let start_time be the time that the liquidation on this margin account began, if it exists

Let timed_out be the boolean representing the type of account:

a. If the liquidation is timed out, then this can be any account.

b. If the liquidation is not timed out, then this must be the liquidator, and it must be a signer.

Check if the entity trying to end the liquidation is not the liquidator.

a. If not, return ErrorCode::UnauthorizedLiquidator.

Record the end of the liquidation.

Emit the LiquidationEnded event for data logging (see table below).

Return Ok(()).

Parameters of liquidate_end.rs:

Name	Description

authority	The pubkey calling the instruction to end liquidation.

margin_account	The account in need of liquidation.

liquidation	The account to persist the state of liquidation.

Events emitted by liquidate_end.rs:

Event Name	Description

events::LiquidationEnded	The event marking the end of liquidation (Includes the margin account pubkey, the authority of the margin account pubkey, and the timed_out boolean that is true if the liquidation has timed out).