Function jet_margin::jet_margin::liquidator_invoke

pub fn liquidator_invoke<'info>(
    ctx: Context<'_, '_, '_, 'info, LiquidatorInvoke<'info>>,
    data: Vec<u8>
) -> Result<()>

Perform an action by invoking another program, for the purposes of liquidating a margin account.

Requires the account already be in the liquidation state, and the signer must be the same liquidator that started the liquidation state.

Accounts

Name	Type	Description

liquidator	signer	The liquidator processing the margin account.

liquidation	writable	The account to persist the state of liquidation.

margin_account	writable	The margin account to proxy an action for.

adapter_program	read_only	The program to be invoked.

adapter_metadata	read_only	The metadata about the proxy program.

Events

Event Name	Description

events::LiquidatorInvokeBegin	Marks the beginning of this liquidation event.

events::PositionEvent (Note that each single event represents an different adapter position)	The PositionEvent describing the change in position.

events::LiquidatorInvokeEnd	Marks the ending of this liquidator event.