Function jet_margin::jet_margin::refresh_position_metadata

pub fn refresh_position_metadata(
    ctx: Context<'_, '_, '_, '_, RefreshPositionMetadata<'_>>
) -> Result<()>

Update the metadata for a position stored in the margin account, in the case where the metadata has changed after the position was created.

Accounts

Name	Type	Description

margin_account	writable	The margin account with the position to be refreshed.

metadata	read_only	The metadata account for the token, which has been updated.

Events

Event Name	Description

events::PositionMetadataRefreshed	Marks the refreshing of position metadata.