Struct jet_margin::AccountPositionList

#[repr(C)]
pub struct AccountPositionList {
    pub length: usize,
    pub map: [AccountPositionKey; 32],
    pub positions: [AccountPosition; 32],
}

Fields
length: usizemap: [AccountPositionKey; 32]positions: [AccountPosition; 32]
Implementations

impl AccountPositionList

pub fn add(
&mut self,
mint: Pubkey
) -> AnchorResult<(AccountPositionKey, &mut AccountPosition)>

Add a position to the position list.

Finds an empty slot in map and positions, and adds an empty position to the slot.

pub fn remove(
&mut self,
mint: &Pubkey,
account: &Pubkey
) -> AnchorResult<AccountPosition>

Remove a position from the margin account.

Error

If an account with the mint does not exist.

If the position’s address is not the same as the account

pub fn get(&self, mint: &Pubkey) -> Option<&AccountPosition>

pub fn get_mut(&mut self, mint: &Pubkey) -> Option<&mut AccountPosition>

pub fn get_key(&self, mint: &Pubkey) -> Option<&AccountPositionKey>