Trait jet_margin::AnchorVerify

pub trait AnchorVerify: Discriminator + Owner {
    fn anchor_verify(info: &AccountInfo<'_>) -> AnchorResult<()> { ... }
}

Execute all the mandatory anchor account verifications that are used during deserialization

performance: don’t have to deserialize (even zero_copy copies)

compatibility: straightforward validation for programs using different anchor versions and non-anchor programs

Provided Methods

fn anchor_verify(info: &AccountInfo<'_>) -> AnchorResult<()>