Function jet_margin::jet_margin::accounting_invoke

pub fn accounting_invoke<'info>(
    ctx: Context<'_, '_, '_, 'info, AccountingInvoke<'info>>,
    data: Vec<u8>
) -> Result<()>

Perform an action by invoking other programs, allowing them only to refresh the state of the margin account to be consistent with the actual underlying prices or positions, but not permitting new position changes.

This instruction does the following:

Emit events::AccountingInvokeBegin event (see Events table below) for data logging.

Invoke the adapter program to update the position for the margin account passed in.

a. For each position that changed, emit each existing adapter position as events::PositionEvent (see Events table below).

Emit events::AccountingInvokeEnd event (see Events table below) for data logging.

Return Ok(()).

Accounts expected with accounting_invoke.rs:

Name	Type	Description

margin_account	writable	The margin account to proxy an action for.

adapter_program	read_only	The program to be invoked.

adapter_metadata	read_only	The metadata about the proxy program.

Events emitted by accounting_invoke.rs:

Name	Description

events::AccountingInvokeBegin	Signify that the accounting invocation process has begun.

events::PositionEvent (Note that each single event represents an different adapter position)	The PositionEvent describing the change in position.

events::AccountingInvokeEnd	Signify that the accounting invocation process has ended.