Function jet_margin::jet_margin::liquidator_invoke

pub fn liquidator_invoke<'info>(
    ctx: Context<'_, '_, '_, 'info, LiquidatorInvoke<'info>>, 
    data: Vec<u8>
) -> Result<()>

Perform an action by invoking another program, for the purposes of liquidating a margin account.

This instruction does the following:

Load the margin account.

Let start_value be the valuation of the margin account before invoking the liquidator.

Emit the LiquidatorInvokeBegin event for data logging (see table below).

Loop through adapter and store positions, getting and storing as margin_account, adapter_program, accounts and signed.

Emit each adapter position as an event (see table below).

Letliquidation be a mutable copy of the liquidated account.

Let end_value be the valuation of the margin account after the liquidation attempt, after verifying that a liquidation did occur.

Emit the LiquidatorInvokeEnd event for data logging (see table below).

Return Ok(()).

Parameters of liquidator_invoke.rs:

Name	Description

liquidator	The liquidator processing the margin account.

liquidation	The account to persist the state of liquidation.

margin_account	The margin account to proxy an action for.

adapter_program	The program to be invoked.

adapter_metadata	The metadata about the proxy program.

Events emitted by liquidator_invoke.rs:

Event Name	Description

events::LiquidatorInvokeBegin	Marks the beginning of this liquidation event (includes the margin account pubkey, the adapter program pubkey, and the liquidator pubkey that is liquidating that margin account or adapter position).

events::PositionEvent (Note that each single event represents an different adapter position)	Each adapter position is emitted as an event (includes the margin account, the adapter program, the accounts, and a value of true for the signed field.

events::LiquidatorInvokeEnd	Marks the ending of this liquidator event (includes the liquidation data and the valuation of the account after liquidation has been performed).